requirements.txt
requests==2.32.5
google-api-python-client==2.185.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
lxml==6.0.1
pytz==2025.2